% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{mymodel}
\alias{mymodel}
\title{TV watching model}
\format{a \code{\link{glm}} object}
\source{
General Social Survey - NORC
}
\usage{
mymodel
}
\description{
Simple example model that predicts hours of watching TV from age and marital status
}

