% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tv}
\alias{tv}
\title{Score the TV watching model}
\usage{
tv(input)
}
\arguments{
\item{input}{data passed on as \code{newdata} to \code{\link{predict}}}
}
\description{
Simple model with two predictor \code{age} (value between 20 and 80) and \code{marital}
("MARRIED", 'DIVORCED", "WIDOWED or "NEVER MARRIED").
}
\examples{
mydata <- data.frame(
   age=c(24, 54, 32, 75),
   marital=c("MARRIED", "DIVORCED", "WIDOWED", "NEVER MARRIED")
)
tv(mydata)
}

