% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{tv_model}
\alias{tv_model}
\title{TV watching model}
\format{a \code{\link{glm}} object}
\source{
General Social Survey - NORC
}
\usage{
tv_model
}
\description{
Simple example model that predicts hours of watching TV from age and marital status
}

