% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.R
\name{tv}
\alias{tv}
\title{Dibuja el histograma de una variable superponiendo la densidad normal ajustada}
\usage{
tv(input)
}
\arguments{
\item{x}{vector de datos cuyo histograma se va a calcular}

\item{dens}{valor lógico: TRUE=Superponer estimador de núcleo de la densidad}
}
\value{
el histograma con la densidad normal superpuesta
}
\description{
Función que dibuja el histograma de una variable x, superponiendo la densidad normal
ajustada. Si el usuario lo desea puede superponer también un estimador de núcleo de la densidad.
}
\examples{
u=rnorm(1000,100,12)
normalHist(u);
normalHist(u,dens=TRUE)
normalHist(u,dens=TRUE,col="lightcyan")

}
