% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.R
\name{tv}
\alias{tv}
\title{Funciones de prueba para predicciones}
\usage{
tv(input)
}
\arguments{
\item{input}{vector de datos}
}
\value{
un vectorde datos de prediccion
}
\description{
Funcion que predice basadndos een una linea de tiwmpo
}
\examples{
newdata <- data.frame(
age = c(24, 54, 32, 75),
marital = c("MARRIED", "DIVORCED", "WIDOWED", "NEVER MARRIED")
)
tv(input = newdata)
}
